fix statistics panel

+++ Assets/Scripts/PlayerStatistics.cs
@@ -5,7 +5,7 @@ namespace Assets.Scripts
     public struct PlayerStatistics
     {
         public Vector2 Coordinates { get; set; }
-        public float Angle { get; set; }
+        public int Angle { get; set; }
         public float Velocity { get; set; }
         public int LaserAmmunitionCount { get; set; }
         public int LaserCooldown { get; set; }
